+++ controllers/podnetworkchaos/tc/tc.go
@@ -16,18 +16,17 @@ package tc
 import (
 	"context"
 	"fmt"
+	"github.com/chaos-mesh/chaos-mesh/pkg/controllerutils"
 
 	v1 "k8s.io/api/core/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/chaos-mesh/chaos-mesh/pkg/chaosdaemon/pb"
-
-	daemonClient "github.com/chaos-mesh/chaos-mesh/pkg/chaosdaemon/client"
 )
 
 // SetTcs makes grpc call to chaosdaemon to flush traffic control rules
 func SetTcs(ctx context.Context, c client.Client, pod *v1.Pod, tcs []*pb.Tc) error {
-	pbClient, err := daemonClient.NewChaosDaemonClient(ctx, c, pod)
+	pbClient, err := controllerutils.NewChaosDaemonClient(ctx, c, pod)
 	if err != nil {
 		return err
 	}
