+++ pkg/chaosctl/common/common.go
@@ -18,6 +18,8 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	grpcUtils "github.com/chaos-mesh/chaos-mesh/pkg/grpc"
+	"github.com/chaos-mesh/chaos-mesh/pkg/mock"
 	"io"
 	"regexp"
 	"strings"
@@ -310,7 +312,7 @@ func GetPidFromPod(ctx context.Context, pod v1.Pod, daemon v1.Pod) (uint32, erro
 	}()
 
 	// TODO: support specify the cert file or get cert file automatically
-	daemonClient, err := daemonClient.NewChaosDaemonClientLocally(int(localPort), "", "", "")
+	daemonClient, err := NewChaosDaemonClientLocally(int(localPort), "", "", "")
 	if err != nil {
 		return 0, errors.Wrapf(err, "failed to create new chaos daemon client with local port %d", localPort)
 	}
@@ -402,4 +404,20 @@ Please check network policy / firewall, or see FAQ on website`, daemon.Name, dae
 
 	}
 	return nil
+}
+
+// NewChaosDaemonClientLocally would create ChaosDaemonClient in localhost
+func NewChaosDaemonClientLocally(port int, caCert string, cert string, key string) (daemonClient.ChaosDaemonClientInterface, error) {
+	if cli := mock.On("MockChaosDaemonClient"); cli != nil {
+		return cli.(daemonClient.ChaosDaemonClientInterface), nil
+	}
+	if err := mock.On("NewChaosDaemonClientError"); err != nil {
+		return nil, err.(error)
+	}
+
+	cc, err := grpcUtils.CreateGrpcConnection("localhost", port, caCert, cert, key)
+	if err != nil {
+		return nil, err
+	}
+	return daemonClient.New(cc), nil
 }
